extract tests for >=3.4
