+++ type-tests/types-min-3.4.test.ts
@@ -1,0 +1,5 @@
+import { testIf, checkDirectory, tsVersion } from './lib'
+
+testIf(tsVersion >= 3.4)('Types >= 3.4', () => {
+  checkDirectory(`${__dirname}/files/min-3.4`)
+})
