+++ type-tests/types.test.ts
@@ -1,4 +1,4 @@
-import { checkDirectory } from 'typings-tester'
+import { checkDirectory } from './lib'
 
 test('Types', () => {
   checkDirectory(`${__dirname}/files`)
